refactor: 타입 분리 및 InputLayout 컴포넌트 공통 컴포넌트로 변경
