+++ src/components/Button/type.ts
@@ -1,0 +1,10 @@
+export type ButtonProps = {
+  text: string;
+  disabled: boolean;
+  type: 'button' | 'submit' | 'reset';
+  handleClick: () => void;
+};
+
+export type ButtonStyleProps = {
+  disabled: boolean;
+};
