+++ src/components/CreditCardCompany/type.ts
@@ -1,0 +1,6 @@
+import * as Type from '@Types/index';
+
+export type CreditCardCompanyProps = {
+  company: Type.CreditCardCompanies;
+  handleClick: () => void;
+};
