+++ src/components/Header/type.ts
@@ -1,0 +1,4 @@
+export type HeaderProps = {
+  title: string;
+  hasBackButton?: boolean;
+};
