+++ src/components/Input/index.tsx
@@ -1,20 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import * as S from './style';
-
-export type InputProps<T> = {
-  type: 'string' | 'number' | 'date' | 'datetime' | 'password';
-  value: T;
-  width: string;
-  textAlign: 'center' | 'start';
-  placeholder?: string;
-  maxLength?: number;
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onClick?: () => void;
-  background?: string;
-  underline?: boolean;
-};
-
-type InputValueType = string | number | readonly string[] | undefined;
+import { InputProps, InputValueType } from './type';
 
 function Input<T extends InputValueType>({
   type,
