+++ src/components/InputLabel/type.ts
@@ -1,0 +1,3 @@
+export type InputLabelProps = {
+  label: string;
+};
