+++ src/components/InputLayout/type.ts
@@ -1,0 +1,4 @@
+export type InputLayoutProps = {
+  children: React.ReactNode;
+  errorMessage: string | null;
+};
