+++ src/pages/CreditCardRegister/CreditCardCVCInput/type.ts
@@ -1,0 +1,5 @@
+export type CreditCardCVCInputProps = {
+  creditCardCVC: string;
+  errorMessage: string | null;
+  updateCVC: (newCVC: string) => void;
+};
