+++ src/pages/CreditCardRegister/CreditCardCompanyModal/CreditCardCompanyModal.tsx
@@ -5,21 +5,17 @@ import * as Type from '@Types/index';
 import CARD_COMPANY from '@Constants/CardCompany';
 
 import * as S from './style';
-
-type CreditCardCompanyModalProps = {
-  updateCompany: (company: Type.CardCompanies) => void;
-  closeModal: () => void;
-};
+import { CreditCardCompanyModalProps } from './type';
 
 function CreditCardCompanyModal({ updateCompany, closeModal }: CreditCardCompanyModalProps) {
-  const handleClickCardCompany = (company: Type.CardCompanies) => {
+  const handleClickCardCompany = (company: Type.CreditCardCompanies) => {
     updateCompany(company);
     closeModal();
   };
 
   return (
     <S.CardCompanyLayout>
-      {(Object.keys(CARD_COMPANY) as Type.CardCompanies[]).map((company) => (
+      {(Object.keys(CARD_COMPANY) as Type.CreditCardCompanies[]).map((company) => (
         <CreditCardCompanyType company={company} key={company} handleClick={() => handleClickCardCompany(company)} />
       ))}
     </S.CardCompanyLayout>
