+++ src/pages/CreditCardRegister/CreditCardCompanyModal/type.ts
@@ -1,0 +1,6 @@
+import * as Type from '@Types/index';
+
+export type CreditCardCompanyModalProps = {
+  updateCompany: (company: Type.CreditCardCompanies) => void;
+  closeModal: () => void;
+};
