+++ src/pages/CreditCardRegister/CreditCardExpiryInput/type.ts
@@ -1,0 +1,5 @@
+export type CreditCardExpiryInputProps = {
+  creditCardExpiry: string;
+  errorMessage: string | null;
+  updateExpiry: (newExpiry: string) => void;
+};
