+++ src/pages/CreditCardRegister/CreditCardNumberInput/type.ts
@@ -1,0 +1,5 @@
+export type CreditCardNumberInputProps = {
+  creditCardNumber: string;
+  errorMessage: string | null;
+  updateNumbers: (numbers: string) => void;
+};
