+++ src/pages/CreditCardRegister/CreditCardOwnerInput/type.ts
@@ -1,0 +1,5 @@
+export type CreditCardOwnerInputProps = {
+  creditCardOwner: string;
+  errorMessage: string | null;
+  updateOwner: (newName: string) => void;
+};
