+++ src/pages/CreditCardRegister/CreditCardPasswordInput/type.ts
@@ -1,0 +1,11 @@
+export type CreditCardPasswordInputProps = {
+  creditCardPassword: {
+    first: string;
+    second: string;
+  };
+  errorMessage: string | null;
+  updatePassword: {
+    first: (newPassword: string) => void;
+    second: (newPassword: string) => void;
+  };
+};
