+++ src/pages/CreditCardRegister/type.ts
@@ -1,0 +1,3 @@
+export type RegisterButtonStyleProps = {
+  disabled: boolean;
+};
