+++ src/styles/common/type.ts
@@ -1,0 +1,3 @@
+export type FlexBoxStyleProps = {
+  justifyContent: 'center' | 'flex-start' | 'flex-end' | 'space-between' | 'space-around';
+};
